use http handler instead of a middleware
